+++ src/main/java/io/github/astrapi69/crypto/key/reader/PemObjectReader.java
@@ -34,6 +34,7 @@ import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.security.PrivateKey;
 import java.security.spec.InvalidKeySpecException;
+import java.util.Optional;
 
 import lombok.NonNull;
 
@@ -70,7 +71,7 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static PemObject getPemObject(final File file) throws IOException
+	public static PemObject getPemObject(final @NonNull File file) throws IOException
 	{
 		PemObject pemObject;
 		try (PemReader pemReader = new PemReader(new InputStreamReader(new FileInputStream(file))))
@@ -89,7 +90,7 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static boolean isPemObject(final File file) throws IOException
+	public static boolean isPemObject(final @NonNull File file) throws IOException
 	{
 		return getPemObject(file) != null;
 	}
@@ -103,7 +104,7 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static Object readPemKeyObject(File keyFile) throws IOException
+	public static Object readPemKeyObject(final @NonNull File keyFile) throws IOException
 	{
 		try (PEMParser pemParser = new PEMParser(
 			new InputStreamReader(new FileInputStream(keyFile))))
@@ -125,7 +126,8 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static PrivateKey readPemPrivateKey(File keyFile, String password) throws IOException
+	public static Optional<PrivateKey> readPemPrivateKey(final @NonNull File keyFile,
+		String password) throws IOException
 	{
 		Object pemKeyObject = readPemKeyObject(keyFile);
 		if (pemKeyObject instanceof PEMEncryptedKeyPair)
@@ -137,20 +139,17 @@ public final class PemObjectReader
 
 			JcaPEMKeyConverter converter = new JcaPEMKeyConverter()
 				.setProvider(SecurityProvider.BC.name());
-			return converter.getPrivateKey(pemKeyPair.getPrivateKeyInfo());
+			return Optional.of(converter.getPrivateKey(pemKeyPair.getPrivateKeyInfo()));
 		}
-		return null;
+		return Optional.empty();
 	}
 
-
 	/**
-	 * Reads the given {@link File}( in *.pem format) that contains a password protected private
-	 * key.
+	 * Reads the given {@link File}( in *.pem format) that contains private key.
 	 *
-	 * @param keyFile
-	 *            the file with the password protected private key
-	 * @return the {@link PrivateKey} object or null if the given file is not a password protected
-	 *         private key.
+	 * @param keyPemFile
+	 *            the file with the private key ( in *.pem format)
+	 * @return the {@link PrivateKey} object or null if the given file is not private key
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 * @throws NoSuchAlgorithmException
@@ -161,23 +160,20 @@ public final class PemObjectReader
 	 *             is thrown if the specified provider is not registered in the security provider
 	 *             list.
 	 */
-	public static PrivateKey readPemPrivateKey(File keyFile) throws IOException,
+	public static PrivateKey readPemPrivateKey(final @NonNull File keyPemFile) throws IOException,
 		NoSuchAlgorithmException, InvalidKeySpecException, NoSuchProviderException
 	{
-		return readPrivateKey(keyFile, KeyPairGeneratorAlgorithm.RSA);
+		return readPrivateKey(keyPemFile, KeyPairGeneratorAlgorithm.RSA);
 	}
 
-
 	/**
-	 * Reads the given {@link File}( in *.pem format) that contains a password protected private
-	 * key.
+	 * Reads the given {@link File}( in *.pem format) that contains private key.
 	 *
-	 * @param keyFile
-	 *            the file with the password protected private key
+	 * @param keyPemFile
+	 *            the file with the private key ( in *.pem format)
 	 * @param algorithm
 	 *            the algorithm for the {@link KeyFactory}
-	 * @return the {@link PrivateKey} object or null if the given file is not a password protected
-	 *         private key.
+	 * @return the {@link PrivateKey} object or null if the given file is not private key
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 * @throws NoSuchAlgorithmException
@@ -188,13 +184,13 @@ public final class PemObjectReader
 	 *             is thrown if the specified provider is not registered in the security provider
 	 *             list.
 	 */
-	public static PrivateKey readPrivateKey(File keyFile, final Algorithm algorithm)
+	public static PrivateKey readPrivateKey(final @NonNull File keyPemFile,
+		final @NonNull Algorithm algorithm)
 		throws IOException, NoSuchAlgorithmException, InvalidKeySpecException,
 		NoSuchProviderException
 	{
-		PemObject pemKeyObject = getPemObject(keyFile);
-		byte[] encoded = pemKeyObject.getContent();
-		return PrivateKeyReader.readPrivateKey(encoded, algorithm.getAlgorithm());
+		return PrivateKeyReader.readPrivateKey(getPemObject(keyPemFile).getContent(),
+			algorithm.getAlgorithm());
 	}
 
 	/**
@@ -206,7 +202,7 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static String toPemFormat(final PemObject pemObject) throws IOException
+	public static String toPemFormat(final @NonNull PemObject pemObject) throws IOException
 	{
 		final StringWriter stringWriter = new StringWriter();
 		final PemWriter pemWriter = new PemWriter(stringWriter);
@@ -224,14 +220,14 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static String toPemFormat(final File file) throws IOException
+	public static Optional<String> toPemFormat(final @NonNull File file) throws IOException
 	{
 		PemObject pemObject = getPemObject(file);
 		if (pemObject != null)
 		{
-			return toPemFormat(pemObject);
+			return Optional.of(toPemFormat(pemObject));
 		}
-		return null;
+		return Optional.empty();
 	}
 
 	/**
@@ -241,7 +237,7 @@ public final class PemObjectReader
 	 *            the pem object
 	 * @return the byte array in the der format
 	 */
-	public static byte[] toDer(final PemObject pemObject)
+	public static byte[] toDer(final @NonNull PemObject pemObject)
 	{
 		return pemObject.getContent();
 	}
@@ -267,7 +263,7 @@ public final class PemObjectReader
 	 * @throws IOException
 	 *             Signals that an I/O exception has occurred.
 	 */
-	public static PemType getPemType(final File file) throws IOException
+	public static PemType getPemType(final @NonNull File file) throws IOException
 	{
 		PemObject pemObject = getPemObject(file);
 		if (pemObject == null)